Make Rust sound work on Linux
An odd bug here: the Rust version plays no audible sound on Linux, with
Pulseaudio.  When I run `strace -f ./target/debug/letters`, it works!  But
with `strace ./target/debug/letters` (no `-f`) it is again silent.

The current fix is to just sleep long enough after enqueueing each sound
for the sound to complete.  That's not great, but I'm going to commit this
mostly-working version.